Fix For Send on Change
The 'lynxSetValue' function should emit both the 'input' and 'change' events. On the 'input' event, validated containers should emit 'lynx-validated' if they do not contain any validated views. On the 'lynx-validated', a submit view with a 'send' value of 'change' should invoke itself if its form's validity state is not 'invalid'.