Fine tuning number.step validation
The number.step validation logic did not take into account significant
digits when calculating validation state. The original calculation
was a simple remainder calculation that would produce inaccurate results
for some floating point situations. For example a `step` of `0.01`
evaluated against a value of `0.34` would result in `invalid` even
though it is a valid step. This change adds logic to use the number of
decimal places in the `step` value to determine significant digits for
evaluating the result of the remainder calculation. If the value has
more decimal digits than the step `invalid` is the result.