Change validation content visibilty before event
This proposed change is to update the visibility of related validation content before dispatching the `lynx-validation-state-change` event. This allows components to make determinations on positioning of the validation content when responding to the `lynx-validation-state-change` event. Without this change, the visibility of the validation content has not been updated so there is not enough information to make those determinations in the `lynx-validation-state-change` event handlers.